패턴을 위한 분리
